Fixed some 10.6 deprecation warnings.

+++ src/jstalkmain.m
@@ -12,7 +12,7 @@
 
 - (void) JSCocoa:(JSCocoaController*)controller hadError:(NSString*)error onLineNumber:(NSInteger)lineNumber atSourceURL:(id)url {
     
-    NSLog(@"Error line %d, %@", lineNumber, error);
+    NSLog(@"Error line %d, %@", (int)lineNumber, error);
     
     exit(1);
 }
